--- a/src/include/common/allocator.h
+++ b/src/include/common/allocator.h
@@ -1,4 +1,5 @@
 #pragma once
+#include <cerrno>
 #include <cstdlib>
 #include "common/typedefs.h"
 namespace terrier::common {
@@ -12,14 +13,16 @@ struct AllocationUtil {
   /**
    * Allocates a chunk of memory whose start address is aligned to the given word size (which is
    * 8 bytes by default) If you ever use an argument that is not 8 bytes, you should really know
-   * what you are doing.
+   * what you are doing. Alignment must be both a power of two and a multiple of sizeof(void *)
    * @param byte_size size of the memory chunk to allocate, in bytes
    * @param alignment the word size to align up to, 8 bytes by default
    * @return allocated memory pointer
    */
-  static byte *AllocateAligned(uint64_t byte_size, uint64_t alignment = sizeof(uint64_t)) {
+  static byte *AllocateAligned(uint64_t byte_size, uint64_t alignment = sizeof(void *)) {
     void *result = nullptr;
-    posix_memalign(&result, alignment, byte_size);
+    int32_t ret UNUSED_ATTRIBUTE = posix_memalign(&result, alignment, byte_size);
+    TERRIER_ASSERT(ret != EINVAL,
+                   "Invalid alignment given to posix_memalign, should be a power of 2 and multiple pf sizeof(void *)");
     return reinterpret_cast<byte *>(result);
   }
 };