--- a/src/include/storage/delta_record.h
+++ b/src/include/storage/delta_record.h
@@ -137,6 +137,10 @@ class PACKED ProjectedRow {
   }
 };
 
+/**
+ * A ProjectedRowInitializer calculates and stores information on how to initialize ProjectedRows
+ * for a specific layout.
+ */
 class ProjectedRowInitializer {
  public:
   /**