--- a/src/include/storage/tuple_access_strategy.h
+++ b/src/include/storage/tuple_access_strategy.h
@@ -39,8 +39,8 @@ class TupleAccessStrategy {
      * @return a pointer to the start of the column. (use as an array)
      */
     byte *ColumnStart(const BlockLayout &layout, const uint16_t col) {
-      return varlen_contents_ +
-             StorageUtil::PadUpToSize(layout.attr_sizes_[col], common::BitmapSize(layout.num_slots_));
+      return StorageUtil::AlignedPtr(layout.attr_sizes_[col],
+                                     varlen_contents_ + common::RawBitmap::SizeInBytes(layout.num_slots_));
     }
 
     /**