--- a/src/include/transaction/transaction_context.h
+++ b/src/include/transaction/transaction_context.h
@@ -67,9 +67,9 @@ class TransactionContext {
    */
   storage::UndoRecord *UndoRecordForInsert(storage::DataTable *const table, const storage::BlockLayout &layout,
                                            const storage::TupleSlot slot) {
-    uint32_t size = storage::UndoRecord::Size(layout, {PRIMARY_KEY_COLUMN_ID});
-    storage::UndoRecord *result = undo_buffer_.NewEntry(size);
-    return storage::UndoRecord::InitializeRecord(result, txn_id_.load(), slot, table, layout, {PRIMARY_KEY_COLUMN_ID});
+    storage::ProjectedRowInitializer initializer(layout, {PRIMARY_KEY_COLUMN_ID});
+    storage::UndoRecord *result = undo_buffer_.NewEntry(storage::UndoRecord::Size(initializer));
+    return storage::UndoRecord::InitializeRecord(result, txn_id_.load(), slot, table, initializer);
   }
 
  private: