this branch vs. master on Single-threaded data table insert on larger tuple (5 cols * 8 bytes), 1 txn and 10M inserts.
```
2018-08-22 16:07:20
Run on (4 X 3100 MHz CPU s)
CPU Caches:
  L1 Data 32K (x2)
  L1 Instruction 32K (x2)
  L2 Unified 262K (x2)
  L3 Unified 4194K (x1)
---------------------------------------------------------------------------------
Benchmark                                          Time           CPU Iterations
---------------------------------------------------------------------------------
DataTableBenchmark/SimpleInsert/real_time       1810 ms       1790 ms          1   5.26979M items/s

2018-08-22 16:09:53
Run on (4 X 3100 MHz CPU s)
CPU Caches:
  L1 Data 32K (x2)
  L1 Instruction 32K (x2)
  L2 Unified 262K (x2)
  L3 Unified 4194K (x1)
---------------------------------------------------------------------------------
Benchmark                                          Time           CPU Iterations
---------------------------------------------------------------------------------
DataTableBenchmark/SimpleInsert/real_time       3439 ms       3392 ms          1   2.77277M items/s
```
Not adding additional benchmarks because we will take a pass through all of them tomorrow anyways