@mbutrovich 2018-08-24
Also are there any debug logging statements that make sense to you to add while you're in the code? I don't think it's a requirement, but just reminding you that it's there.